criados metodos findAll e FindById no controller de produto

+++ AspNetShop.ProductAPI/Repository/ProductRepository.cs
@@ -26,7 +26,7 @@ public class ProductRepository : IProductRepository
     public async Task<ProductVO> FindById(long id)
     {
         Product product = await _context.Products
-            .Where(p => p.Id == id).FirstOrDefaultAsync() ?? throw new InvalidOperationException();
+            .Where(p => p.Id == id).FirstOrDefaultAsync();
         return _mapper.Map<ProductVO>(product);
     }
 
@@ -51,7 +51,8 @@ public class ProductRepository : IProductRepository
         try
         {
             Product product = await _context.Products
-                .Where(p => p.Id == id).FirstOrDefaultAsync() ?? throw new InvalidOperationException();
+                .Where(p => p.Id == id).FirstOrDefaultAsync();
+            if (product == null) return false;
             _context.Products.Remove(product);
             await _context.SaveChangesAsync();
             return true;
